Correct improper scaling of WaveformAtAPointFT
When converting from time-domain to frequency-domain, Numerical Recipes (like many FFT routines) outputs an ugly normalization.  To make it coincide with the usual physicist's idea of an FT, we need to multiply by the time step.  This time step was in the incorrect units -- which wasn't a problem for anything that didn't use (or would cancel out) the amplitude of the FT.  The amplitude still needs to be adjusted to remove the r/M in, e.g., r*h/M.