Speed up extrapolation a lot
Because of the whole SWIG fiasco, Waveform methods that return
Waveform objects actually return _Waveform objects, which need to be
translated to Waveform objects.  This takes in the neighborhood of
0.01 seconds per operation -- not a problem for the occasional usage,
but really bad when you call the methods tens of millions of times,
which _Extrapolate was doing.  I've changed the code to store the data
in intermediate numpy arrays, which have fast access.  Then, once
everything is done computing, I just copy those numpy arrays into the
Waveform objects.  It still goes through the _Waveform conversion, but
only ~4 times, so that's not a problem.  In particular, extrapolation
now takes around 4 minutes, rather than the weeks it might have taken
before.  About half of this time is still due to that conversion
process (in other routines), but I can live with it for now.