Waveforms.cpp: fix bug in Differentiate
Previously, this function would return an UnknownDataType regardless
of the input data type. It now correctly returns the data type
resulting from the derivative of the input. If the input is anything
other than h or hdot, the result is UnknownDataType.

Also removed a warning about negating the real part when taking
a derivative that results in Psi4. Since h and Psi4 have the
same spin weight, no conjugation of \ddot{h} should be needed
to get agreement with Psi4 (according to Mike).